Revert "Reapply "fix(mapper): added converter to service impl""
This reverts commit 8fc513b78d64961d352167cd7a5cc4a76cec46b9.

+++ src/main/java/com/gearx/feature/brand/service/Impl/BrandServiceImpl.java
@@ -68,7 +68,7 @@ public class BrandServiceImpl implements BrandService {
         params.put("limit", limit);
 
         List<Brand> rows = brandMapper.pageSearch(params);
-        int total = brandMapper.countPageSearch(params);
+        int total = (int) brandMapper.countPageSearch(params);
 
         return brandConverter.toResponsePage(rows, offset, limit, total);
     }
